Another unsuccessful try at a solution...
