+++ src/main/resources/data/planet-b/worldgen/biome/plains.json
@@ -1,0 +1,33 @@
+{
+  "precipitation": "rain",
+  "category": "none",
+  "temperature": 0.3,
+  "temperature_modifier": "none",
+  "downfall": 0.2,
+  "creature_spawn_probability": 0,
+  "effects": {
+    "sky_color": 3357900,
+    "grass_color": 3386026,
+    "foliage_color": 5603498,
+    "fog_color": 3356262,
+    "water_color": 819,
+    "water_fog_color": 0
+  },
+  "carvers": {},
+  "spawners": {},
+  "spawn_costs": {},
+  "features": [
+    [],
+    [],
+    [],
+    [],
+    [],
+    [],
+    [],
+    [],
+    [],
+    [],
+    ["techreborn:rubber_tree"],
+    []
+  ]
+}
