--- src/main/resources/data/planet-b/worldgen/structure_set/sparse_vegetation.json
@@ -1,18 +1,0 @@
-{
-  "structures": [
-    {
-      "structure": "planet-b:single_tree",
-      "weight": 24
-    },
-    {
-      "structure": "planet-b:wither_rose",
-      "weight": 1
-    }
-  ],
-  "placement": {
-    "type": "minecraft:random_spread",
-    "spacing": 8,
-    "separation": 5,
-    "salt": 8192329
-  }
-}
